@@ -1,23 +1,29 @@
 import { TopBar } from "@equinor/eds-core-react";
-import { createStyles, makeStyles } from "@material-ui/core";
+import { styled } from "@mui/material/styles";
 import React from "react";
 import DateTimeSlider from "./DateTimeSlider";
 import FlowRateSelector from "./FlowRateSelector";
 import NodeInfoSelector from "./NodeInfoSelector";
 import { DataInfos } from "../../redux/types";
 
-const useStyles = makeStyles(() =>
-    createStyles({
-        topBar: {
-            minHeight: "90px",
-        },
-        actions: {
-            position: "relative",
-            display: "flex",
-            flexDirection: "row",
-        },
-    })
-);
+const PREFIX = "SettingsBar";
+
+const classes = {
+    topBar: `${PREFIX}-topBar`,
+    actions: `${PREFIX}-actions`,
+};
+
+const StyledTopBar = styled(TopBar)(() => ({
+    [`&.${classes.topBar}`]: {
+        minHeight: "90px",
+    },
+
+    [`& .${classes.actions}`]: {
+        position: "relative",
+        display: "flex",
+        flexDirection: "row",
+    },
+}));
 
 interface Props {
     edge_options: DataInfos;
@@ -26,9 +32,8 @@ interface Props {
 
 const SettingsBar: React.FC<Props> = React.memo(
     ({ edge_options, node_options }: Props) => {
-        const classes = useStyles();
         return (
-            <TopBar className={classes.topBar}>
+            <StyledTopBar className={classes.topBar}>
                 <TopBar.Header className={classes.actions}>
                     <FlowRateSelector edge_options={edge_options} />
                     <NodeInfoSelector node_options={node_options} />
@@ -36,7 +41,7 @@ const SettingsBar: React.FC<Props> = React.memo(
                 <TopBar.Actions className={classes.actions}>
                     <DateTimeSlider />
                 </TopBar.Actions>
-            </TopBar>
+            </StyledTopBar>
         );
     }
 );
