@@ -1,4 +1,4 @@
-import { createStyles, makeStyles, Theme } from "@material-ui/core";
+import { styled } from "@mui/material/styles";
 import React, { useMemo } from "react";
 import { useResizeDetector } from "react-resize-detector";
 import { PlotData } from "../../utils/dataUtil";
@@ -8,16 +8,13 @@ import { getLayout, Padding } from "./plotUtil";
 import StratigraphyPlot from "./StratigraphyPlot";
 import WellsPlot from "./WellsPlot";
 
-const useStyles = makeStyles((theme: Theme) =>
-    createStyles({
-        root: {
-            display: "flex",
-            flex: 1,
-            height: "80%",
-            padding: theme.spacing(1),
-        },
-    })
-);
+const Root = styled("div")(({ theme }) => ({
+    display: "flex",
+    flex: 1,
+    height: "80%",
+    padding: theme.spacing(1),
+}));
+
 interface Props {
     timeSteps: string[];
     plotData: PlotData;
@@ -28,7 +25,6 @@ const padding: Padding = { left: 80, right: 50, top: 70, bottom: 50 };
 /* eslint-disable react/prop-types */
 const WellCompletionsPlot: React.FC<Props> = React.memo(
     ({ timeSteps, plotData }) => {
-        const classes = useStyles();
         const { width, height, ref } = useResizeDetector({
             refreshMode: "debounce",
             refreshRate: 50,
@@ -44,8 +40,7 @@ const WellCompletionsPlot: React.FC<Props> = React.memo(
 
         return (
             <TooltipProvider>
-                <div
-                    className={classes.root}
+                <Root
                     ref={ref as React.LegacyRef<HTMLDivElement>}
                     data-tip
                     data-for="plot-tooltip"
@@ -75,7 +70,7 @@ const WellCompletionsPlot: React.FC<Props> = React.memo(
                             />
                         </svg>
                     )}
-                </div>
+                </Root>
             </TooltipProvider>
         );
     }
