@@ -1,25 +1,29 @@
 import { NativeSelect } from "@equinor/eds-core-react";
-import { createStyles, makeStyles, Theme } from "@material-ui/core";
-import React, { useCallback } from "react";
+import { styled } from "@mui/material/styles";
+import React, { ChangeEvent, useCallback } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { updateTimeAggregation } from "../../redux/actions";
 import { WellCompletionsState } from "../../redux/store";
 import { TimeAggregations } from "../../redux/types";
 
-const useStyles = makeStyles((theme: Theme) =>
-    createStyles({
-        root: {
-            minWidth: "170px",
-            maxWidth: "170px",
-            padding: theme.spacing(1),
-        },
-    })
-);
+const PREFIX = "TimeAggregationSelector";
+
+const classes = {
+    root: `${PREFIX}-root`,
+};
+
+const StyledNativeSelect = styled(NativeSelect)(({ theme }) => ({
+    [`&.${classes.root}`]: {
+        minWidth: "170px",
+        maxWidth: "170px",
+        padding: theme.spacing(1),
+    },
+}));
+
 /**
  * A dropdown for selecting the time aggregation mode
  */
 const TimeAggregationSelector: React.FC = React.memo(() => {
-    const classes = useStyles();
     // Redux
     const dispatch = useDispatch();
     const timeAggregation = useSelector(
@@ -27,12 +31,14 @@ const TimeAggregationSelector: React.FC = React.memo(() => {
     );
     // Handlers
     const handleSelectedItemChange = useCallback(
-        (event) => dispatch(updateTimeAggregation(event.target.value)),
+        (event: ChangeEvent<HTMLSelectElement>) =>
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            dispatch(updateTimeAggregation(event.target.value as any)),
         [dispatch]
     );
     //Render
     return (
-        <NativeSelect
+        <StyledNativeSelect
             className={classes.root}
             id="time-aggregation-selector"
             label="Time Aggregation"
@@ -42,7 +48,7 @@ const TimeAggregationSelector: React.FC = React.memo(() => {
             {Object.keys(TimeAggregations).map((mode) => (
                 <option key={mode}>{mode}</option>
             ))}
-        </NativeSelect>
+        </StyledNativeSelect>
     );
 });
 
