@@ -1,25 +1,29 @@
 /* eslint-disable react-hooks/exhaustive-deps */ // remove when ready to fix these.
 
 import { Search } from "@equinor/eds-core-react";
-import { createStyles, makeStyles, Theme } from "@material-ui/core";
+import { styled } from "@mui/material/styles";
 import { throttle } from "lodash";
 import React, { useCallback } from "react";
 import { useDispatch } from "react-redux";
 import { updateWellSearchText } from "../../redux/actions";
 
-const useStyles = makeStyles((theme: Theme) =>
-    createStyles({
-        root: {
-            padding: theme.spacing(1),
-            maxWidth: "250px",
-        },
-    })
-);
+const PREFIX = "WellFilter";
+
+const classes = {
+    root: `${PREFIX}-root`,
+};
+
+const Root = styled("div")(({ theme }) => ({
+    [`&.${classes.root}`]: {
+        padding: theme.spacing(1),
+        maxWidth: "250px",
+    },
+}));
+
 /**
  * A search textfield to search wells by their names
  */
 const WellFilter: React.FC = React.memo(() => {
-    const classes = useStyles();
     // Redux
     const dispatch = useDispatch();
     // Handlers
@@ -37,14 +41,14 @@ const WellFilter: React.FC = React.memo(() => {
     );
 
     return (
-        <div className={classes.root}>
+        <Root className={classes.root}>
             <Search
                 aria-label="sitewide"
                 id="search-well-name"
                 placeholder="Search well names"
                 onChange={onChange}
             />
-        </div>
+        </Root>
     );
 });
 
