@@ -1,24 +1,28 @@
 import { NativeSelect } from "@equinor/eds-core-react";
-import { createStyles, makeStyles, Theme } from "@material-ui/core";
+import { styled } from "@mui/material/styles";
 import React, { useCallback } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { updateWellsPerPage } from "../../redux/actions";
 import { WellCompletionsState } from "../../redux/store";
 
+const PREFIX = "WellsPerPageSelector";
+
+const classes = {
+    root: `${PREFIX}-root`,
+};
+
+const StyledNativeSelect = styled(NativeSelect)(({ theme }) => ({
+    [`&.${classes.root}`]: {
+        padding: theme.spacing(1),
+        maxWidth: "170px",
+    },
+}));
+
 const wellsPerPageOptions = [10, 25, 50];
-const useStyles = makeStyles((theme: Theme) =>
-    createStyles({
-        root: {
-            padding: theme.spacing(1),
-            maxWidth: "170px",
-        },
-    })
-);
 /**
  * A drop down for selecting how many wells to display per page
  */
 const WellsPerPageSelector: React.FC = React.memo(() => {
-    const classes = useStyles();
     // Redux
     const dispatch = useDispatch();
     const wellsPerPage = useSelector(
@@ -26,12 +30,13 @@ const WellsPerPageSelector: React.FC = React.memo(() => {
     );
     // Handlers
     const onWellsPerPageChange = useCallback(
-        (event) => dispatch(updateWellsPerPage(event.target.value)),
+        (event: React.ChangeEvent<HTMLSelectElement>) =>
+            dispatch(updateWellsPerPage(+event.target.value)),
         [dispatch]
     );
     // Render
     return (
-        <NativeSelect
+        <StyledNativeSelect
             label={"Wells per page"}
             id="wells-per-page-select"
             className={classes.root}
@@ -43,7 +48,7 @@ const WellsPerPageSelector: React.FC = React.memo(() => {
                     {value}
                 </option>
             ))}
-        </NativeSelect>
+        </StyledNativeSelect>
     );
 });
 
