@@ -1,11 +1,5 @@
-import {
-    createStyles,
-    Divider,
-    Drawer,
-    makeStyles,
-    // eslint-disable-next-line prettier/prettier
-    Theme
-} from "@material-ui/core";
+import { Divider, Drawer } from "@mui/material";
+import { styled } from "@mui/material/styles";
 import clsx from "clsx";
 import React, { useContext, useMemo } from "react";
 import { useSelector } from "react-redux";
@@ -21,65 +15,82 @@ import WellFilter from "./Settings/WellFilter";
 import WellPagination from "./Settings/WellPagination";
 import ZoneSelector from "./Settings/ZoneSelector";
 
+const PREFIX = "WellCompletionsViewer";
+
+const classes = {
+    root: `${PREFIX}-root`,
+    main: `${PREFIX}-main`,
+    drawer: `${PREFIX}-drawer`,
+    drawerShift: `${PREFIX}-drawerShift`,
+    drawerPaper: `${PREFIX}-drawerPaper`,
+    drawerHeader: `${PREFIX}-drawerHeader`,
+    content: `${PREFIX}-content`,
+    contentShift: `${PREFIX}-contentShift`,
+};
+
+const Root = styled("div")(({ theme }) => ({
+    [`& .${classes.root}`]: {
+        position: "relative",
+        display: "flex",
+        flex: 1,
+        flexDirection: "column",
+        width: "100%",
+        height: "90%",
+    },
+
+    [`& .${classes.main}`]: {
+        position: "relative",
+        display: "flex",
+        flex: 1,
+        flexDirection: "row",
+        height: "100%",
+    },
+
+    [`& .${classes.drawer}`]: {
+        zIndex: 0,
+        width: drawerWidth,
+        flexShrink: 0,
+    },
+
+    [`& .${classes.drawerShift}`]: {
+        width: 0,
+        flexShrink: 0,
+        display: "none",
+    },
+
+    [`& .${classes.drawerPaper}`]: {
+        position: "relative",
+    },
+
+    [`& .${classes.drawerHeader}`]: {
+        display: "flex",
+        alignItems: "center",
+        padding: theme.spacing(0, 1),
+        // necessary for content to be below app bar
+        ...theme.mixins.toolbar,
+        justifyContent: "flex-start",
+    },
+
+    [`& .${classes.content}`]: {
+        width: "100%",
+        transition: theme.transitions.create("margin", {
+            easing: theme.transitions.easing.sharp,
+            duration: theme.transitions.duration.leavingScreen,
+        }),
+    },
+
+    [`& .${classes.contentShift}`]: {
+        width: `calc(100% - ${drawerWidth}px)`,
+        transition: theme.transitions.create("margin", {
+            easing: theme.transitions.easing.easeOut,
+            duration: theme.transitions.duration.enteringScreen,
+        }),
+    },
+}));
+
 const drawerWidth = 270;
-const useStyles = makeStyles((theme: Theme) =>
-    createStyles({
-        root: {
-            position: "relative",
-            display: "flex",
-            flex: 1,
-            flexDirection: "column",
-            width: "100%",
-            height: "90%",
-        },
-        main: {
-            position: "relative",
-            display: "flex",
-            flex: 1,
-            flexDirection: "row",
-            height: "100%",
-        },
-        drawer: {
-            zIndex: 0,
-            width: drawerWidth,
-            flexShrink: 0,
-        },
-        drawerShift: {
-            width: 0,
-            flexShrink: 0,
-            display: "none",
-        },
-        drawerPaper: {
-            position: "relative",
-        },
-        drawerHeader: {
-            display: "flex",
-            alignItems: "center",
-            padding: theme.spacing(0, 1),
-            // necessary for content to be below app bar
-            ...theme.mixins.toolbar,
-            justifyContent: "flex-start",
-        },
-        content: {
-            width: "100%",
-            transition: theme.transitions.create("margin", {
-                easing: theme.transitions.easing.sharp,
-                duration: theme.transitions.duration.leavingScreen,
-            }),
-        },
-        contentShift: {
-            width: `calc(100% - ${drawerWidth}px)`,
-            transition: theme.transitions.create("margin", {
-                easing: theme.transitions.easing.easeOut,
-                duration: theme.transitions.duration.enteringScreen,
-            }),
-        },
-    })
-);
 
 const WellCompletionsViewer: React.FC = () => {
-    const classes = useStyles();
-
     // Use input data directly
     const data = useContext(DataContext);
     // Create plot data with the selected time step(s)
@@ -117,7 +128,7 @@ const WellCompletionsViewer: React.FC = () => {
     if (!data) return <div />;
     // Render
     return (
-        <div className={classes.root}>
+        <Root className={classes.root}>
             {/* We detect the resize of the element and resize the plot accordingly */}
             <ReactResizeDetector handleWidth handleHeight>
                 {({ width }) => (
@@ -170,7 +181,7 @@ const WellCompletionsViewer: React.FC = () => {
                     </>
                 )}
             </ReactResizeDetector>
-        </div>
+        </Root>
     );
 };
 
