@@ -20,7 +20,7 @@ import "./commands";
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
 
-import { mount } from "@cypress/react";
+import { mount } from "@cypress/react18";
 
 // Augment the Cypress namespace to include type definitions for
 // your custom command.
