@@ -37,7 +37,7 @@ import { cloneDeep } from "lodash";
 import { colorTables } from "@emerson-eps/color-tables";
 import { getModelMatrixScale } from "../layers/utils/layerTools";
 import { OrbitController, OrthographicController } from "@deck.gl/core/typed";
-import { MjolnirEvent } from "mjolnir.js";
+import { MjolnirEvent, MjolnirPointerEvent } from "mjolnir.js";
 import IntersectionView from "../views/intersectionView";
 
 type BoundingBox = [number, number, number, number, number, number];
@@ -342,10 +342,6 @@ const Map: React.FC<MapProps> = ({
     const [viewsProps, setViewsProps] = useState<ViewportType[]>([]);
     const [alteredLayers, setAlteredLayers] = useState<LayersList>([]);
 
-    const [centerOfData, setCenterOfData] = useState<[number, number, number]>([
-        0, 0, 0,
-    ]);
-
     const [viewPortMargins, setViewPortMargins] = useState<marginsType>({
         left: 0,
         right: 0,
@@ -364,7 +360,7 @@ const Map: React.FC<MapProps> = ({
     const initialViewState = getViewState(
         viewPortMargins,
         boundsInitial,
-        centerOfData,
+        boundingBoxCenter(reportedBoundingBoxAcc),
         views?.viewports?.[0].target,
         views?.viewports?.[0].zoom,
         deckRef.current?.deck
@@ -399,13 +395,10 @@ const Map: React.FC<MapProps> = ({
                           deckRef.current?.deck
                       );
 
-                const minZoom = is3D ? -12 : -15;
-                const maxZoom = is3D ? +12 : +15;
+                const minZoom = is3D ? -12 : -12;
+                const maxZoom = is3D ? +12 : +4;
 
-                return [
-                    item.id,
-                    { ...viewState, minZoom: minZoom, maxZoom: maxZoom },
-                ];
+                return [item.id, { ...viewState, minZoom, maxZoom }];
             });
 
             const tempViewStates = Object.fromEntries(viewStateMap);
@@ -433,22 +426,30 @@ const Map: React.FC<MapProps> = ({
                 viewsProps.map((item) => [item.id, cameraPosition])
             ) as unknown as Record<string, ViewStateType>;
         } else {
+            const isBoundsDefined = typeof bounds !== "undefined";
+
             tempViewStates = Object.fromEntries(
                 viewsProps.map((item, index) => {
-                    const viewState = viewStates[item.id];
-                    return [
-                        item.id,
-                        typeof viewState !== "undefined"
-                            ? viewState
-                            : getViewState(
+                    let viewState = viewStates[item.id];
+                    if (typeof viewState === "undefined") {
+                        viewState = isBoundsDefined
+                            ? getViewState(
                                   viewPortMargins,
                                   boundsInitial,
-                                  centerOfData,
+                                  boundingBoxCenter(reportedBoundingBoxAcc),
                                   views?.viewports?.[index].target,
                                   views?.viewports?.[index].zoom,
                                   deckRef.current?.deck
-                              ),
-                    ];
+                              )
+                            : getViewState3D(
+                                  views?.viewports?.[index]?.show3D ?? false,
+                                  reportedBoundingBoxAcc,
+                                  views?.viewports?.[index].zoom,
+                                  deckRef.current?.deck
+                              );
+                    }
+
+                    return [item.id, viewState];
                 })
             );
         }
@@ -460,7 +461,6 @@ const Map: React.FC<MapProps> = ({
     }, [
         boundsInitial,
         cameraPosition,
-        centerOfData,
         isCameraPositionDefined,
         viewPortMargins,
         views?.viewports,
@@ -481,7 +481,7 @@ const Map: React.FC<MapProps> = ({
                     getViewState(
                         viewPortMargins,
                         boundsInitial,
-                        centerOfData,
+                        boundingBoxCenter(reportedBoundingBoxAcc),
                         views?.viewports?.[index].target,
                         views?.viewports?.[index].zoom,
                         deckRef.current?.deck
@@ -499,7 +499,7 @@ const Map: React.FC<MapProps> = ({
         cameraPosition,
         viewPortMargins,
         boundsInitial,
-        centerOfData,
+        reportedBoundingBoxAcc,
         views?.viewports,
     ]);
 
@@ -517,34 +517,38 @@ const Map: React.FC<MapProps> = ({
         );
         setReportedBoundingBoxAcc(union_of_reported_bboxes);
 
-        const center = boundingBoxCenter(
-            union_of_reported_bboxes // note this may include axesLayer
-        );
-        setCenterOfData(center);
-
         // If "bounds" or "cameraPosition" is not defined "viewState" will be
         // calculated based on the union of the reported bounding boxes from each layer.
         if (!didUserChangeCamera && !isCameraPositionDefined) {
             calcDefaultViewStates(union_of_reported_bboxes);
         }
         // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [didUserChangeCamera, isCameraPositionDefined, reportedBoundingBox]);
+    }, [reportedBoundingBox]);
 
     // react on bounds prop change
     useEffect(() => {
+        const isBoundsDefined = typeof bounds !== "undefined";
+
         let tempViewStates: Record<string, ViewStateType> = {};
         if (!isCameraPositionDefined) {
             tempViewStates = Object.fromEntries(
                 viewsProps.map((item, index) => [
                     item.id,
-                    getViewState(
-                        viewPortMargins,
-                        boundsInitial,
-                        centerOfData,
-                        views?.viewports?.[index].target,
-                        views?.viewports?.[index].zoom,
-                        deckRef.current?.deck
-                    ),
+                    isBoundsDefined
+                        ? getViewState(
+                              viewPortMargins,
+                              boundsInitial,
+                              boundingBoxCenter(reportedBoundingBoxAcc),
+                              views?.viewports?.[index].target,
+                              views?.viewports?.[index].zoom,
+                              deckRef.current?.deck
+                          )
+                        : getViewState3D(
+                              views?.viewports?.[index]?.show3D ?? false,
+                              reportedBoundingBoxAcc,
+                              views?.viewports?.[index].zoom,
+                              deckRef.current?.deck
+                          ),
                 ])
             );
             if (viewsProps[0] !== undefined) {
@@ -553,14 +557,7 @@ const Map: React.FC<MapProps> = ({
             setViewStates(tempViewStates);
         }
         // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [
-        bounds,
-        boundsInitial,
-        isCameraPositionDefined,
-        viewPortMargins,
-        views?.viewports,
-        viewsProps,
-    ]);
+    }, [boundsInitial, cameraPosition, views?.viewports, viewsProps]);
 
     // react on cameraPosition prop change
     useEffect(() => {
@@ -580,7 +577,8 @@ const Map: React.FC<MapProps> = ({
             );
             setViewStates(tempViewStates);
         }
-    }, [cameraPosition, initialViewState, isCameraPositionDefined, viewsProps]);
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [cameraPosition, viewsProps]);
 
     // Used for scaling in z direction using arrow keys.
     const [scaleZ, setScaleZ] = useState<number>(1);
@@ -622,7 +620,7 @@ const Map: React.FC<MapProps> = ({
             calcDefaultViewStates(reportedBoundingBoxAcc);
         }
         // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [bounds, views]);
+    }, [views]);
 
     useEffect(() => {
         if (layers == undefined) return;
@@ -788,7 +786,7 @@ const Map: React.FC<MapProps> = ({
         (
             type: "click" | "hover",
             infos: PickingInfo[],
-            event: Record<string, unknown>
+            event: MjolnirEvent
         ): void => {
             if (!onMouseEvent) return;
             const ev = handleMouseEvent(type, infos, event);
@@ -800,7 +798,7 @@ const Map: React.FC<MapProps> = ({
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const [hoverInfo, setHoverInfo] = useState<any>([]);
     const onHover = useCallback(
-        (pickInfo, event) => {
+        (pickInfo: PickingInfo, event: MjolnirEvent) => {
             const infos = getPickingInfos(pickInfo, event);
             setHoverInfo(infos); //  for InfoCard pickInfos
             callOnMouseEvent?.("hover", infos, event);
@@ -809,7 +807,7 @@ const Map: React.FC<MapProps> = ({
     );
 
     const onClick = useCallback(
-        (pickInfo, event) => {
+        (pickInfo: PickingInfo, event: MjolnirEvent) => {
             const infos = getPickingInfos(pickInfo, event);
             callOnMouseEvent?.("click", infos, event);
         },
@@ -868,7 +866,8 @@ const Map: React.FC<MapProps> = ({
     );
 
     const onViewStateChange = useCallback(
-        ({ viewId, viewState }) => {
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        ({ viewId, viewState }: { viewId: string; viewState: any }) => {
             const viewports = views?.viewports || [];
             const isSyncIds = viewports
                 .filter((item) => item.isSync)
@@ -911,7 +910,8 @@ const Map: React.FC<MapProps> = ({
                 views={deckGLViews}
                 layerFilter={layerFilter}
                 layers={deckGLLayers}
-                // @ts-expect-error this prop doesn't exists directly on DeckGL, but on Deck.Context
+                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                // @ts-expect-error
                 userData={{
                     setEditedData: (updated_prop: Record<string, unknown>) => {
                         setSelectedWell(updated_prop["selectedWell"] as string);
@@ -1199,7 +1199,7 @@ function createViews(
                 width: "100%",
                 height: "100%",
                 flipY: false,
-                far: 99999,
+                far: +99999,
                 near: -99999,
             })
         );
@@ -1293,14 +1293,14 @@ function getViews(views: ViewsType | undefined): ViewportType[] {
 function handleMouseEvent(
     type: "click" | "hover",
     infos: PickingInfo[],
-    event: Record<string, unknown>
+    event: MjolnirEvent
 ) {
     const ev: MapMouseEvent = {
         type: type,
         infos: infos,
     };
     if (ev.type === "click") {
-        if (event["rightButton"]) ev.type = "contextmenu";
+        if ((event as MjolnirPointerEvent).rightButton) ev.type = "contextmenu";
     }
     for (const info of infos as LayerPickInfo[]) {
         if (info.coordinate) {
