@@ -171,7 +171,8 @@ export default class Axes2DLayer extends Layer<Axes2DLayerProps<unknown>> {
         min: number,
         max: number,
         viewSide: ViewSide,
-        pixel2world: number
+        pixel2world: number,
+        zoomScale: number
     ): [number[], LabelData[]] {
         const ndecimals = 0;
         const n_minor_ticks = 3;
@@ -212,11 +213,11 @@ export default class Axes2DLayer extends Layer<Axes2DLayerProps<unknown>> {
         const isHorizontal =
             viewSide === ViewSide.Top || viewSide === ViewSide.Bottom;
 
-        const ticks = GetTicks(min, max, L); // Note: this may be replaced by NiceTicks npm package.  // XXX RENAME
+        const ticks = GetTicks(min, max, L); // Note: this may be replaced by NiceTicks npm package.
 
         // z value of all lines and labels. In camera/view coordinates. This
         // ensures lines will be closer to camera than rest of model.
-        const z_tick = isHorizontal ? -100 : -110; // horizontal rulers in front.
+        const z_depth = 99 * (2 / zoomScale);
 
         const tick_length =
             viewSide === ViewSide.Left || viewSide === ViewSide.Bottom
@@ -231,11 +232,11 @@ export default class Axes2DLayer extends Layer<Axes2DLayerProps<unknown>> {
 
             // tick line start
             if (isHorizontal) {
-                lines.push(tick, y_tick, z_tick); // tick line start
-                lines.push(tick, y_tick + tick_length, z_tick); // tick line end.
+                lines.push(tick, y_tick, z_depth); // tick line start
+                lines.push(tick, y_tick + tick_length, z_depth); // tick line end.
             } else {
-                lines.push(x_tick, tick, z_tick);
-                lines.push(x_tick + tick_length, tick, z_tick);
+                lines.push(x_tick, tick, z_depth);
+                lines.push(x_tick + tick_length, tick, z_depth);
             }
         }
 
@@ -254,11 +255,11 @@ export default class Axes2DLayer extends Layer<Axes2DLayerProps<unknown>> {
                 i++;
 
                 if (isHorizontal) {
-                    lines.push(tick, y_tick, z_tick); // tick line start
-                    lines.push(tick, y_tick + 0.5 * tick_length, z_tick); // tick line end.
+                    lines.push(tick, y_tick, z_depth); // tick line start
+                    lines.push(tick, y_tick + 0.5 * tick_length, z_depth); // tick line end.
                 } else {
-                    lines.push(x_tick, tick, z_tick);
-                    lines.push(x_tick + 0.5 * tick_length, tick, z_tick);
+                    lines.push(x_tick, tick, z_depth);
+                    lines.push(x_tick + 0.5 * tick_length, tick, z_depth);
                 }
             }
 
@@ -270,11 +271,11 @@ export default class Axes2DLayer extends Layer<Axes2DLayerProps<unknown>> {
                 i++;
 
                 if (isHorizontal) {
-                    lines.push(tick, y_tick, z_tick);
-                    lines.push(tick, y_tick + 0.5 * tick_length, z_tick);
+                    lines.push(tick, y_tick, z_depth);
+                    lines.push(tick, y_tick + 0.5 * tick_length, z_depth);
                 } else {
-                    lines.push(x_tick, tick, z_tick);
-                    lines.push(x_tick + 0.5 * tick_length, tick, z_tick);
+                    lines.push(x_tick, tick, z_depth);
+                    lines.push(x_tick + 0.5 * tick_length, tick, z_depth);
                 }
             }
         }
@@ -310,11 +311,11 @@ export default class Axes2DLayer extends Layer<Axes2DLayerProps<unknown>> {
         const p4_v = word2view(viewMatrix, p4);
 
         // Distance camera background in view space.
-        const z_dist = 101;
-        p1_v[2] = -z_dist;
-        p2_v[2] = -z_dist;
-        p3_v[2] = -z_dist;
-        p4_v[2] = -z_dist;
+        const z_dist = 0;
+        p1_v[2] = z_dist;
+        p2_v[2] = z_dist;
+        p3_v[2] = z_dist;
+        p4_v[2] = z_dist;
 
         /*eslint-disable */
         const background_lines: number[] = [ 
@@ -352,11 +353,11 @@ export default class Axes2DLayer extends Layer<Axes2DLayerProps<unknown>> {
         const p4_v = word2view(viewMatrix, p4);
 
         // Distance camera background in view space.
-        const z_dist = 105;
-        p1_v[2] = -z_dist;
-        p2_v[2] = -z_dist;
-        p3_v[2] = -z_dist;
-        p4_v[2] = -z_dist;
+        const z_dist = 0;
+        p1_v[2] = z_dist;
+        p2_v[2] = z_dist;
+        p3_v[2] = z_dist;
+        p4_v[2] = z_dist;
 
         /*eslint-disable */
         const background_lines: number[] = [ 
@@ -483,6 +484,8 @@ export default class Axes2DLayer extends Layer<Axes2DLayerProps<unknown>> {
         const pixel2world = Math.sqrt(v[0] * v[0] + v[1] * v[1]) / 100;
 
         const { viewMatrix } = this.context.viewport;
+        const zoom = this.context.viewport.zoom;
+        const zoomScale = Math.pow(2, zoom);
 
         const mh = this.props.marginH * pixel2world;
         const mv = this.props.marginV * pixel2world;
@@ -497,14 +500,17 @@ export default class Axes2DLayer extends Layer<Axes2DLayerProps<unknown>> {
         let background_lines: number[] = [];
         let labelData: LabelData[] = [];
 
+        const zDepth = 2 / zoomScale;
+
         //- BOTTOM RULER ----------------------------------------
         if (this.props.isBottomRuler) {
-            const axes = [xMin, yMin + mv, 0, xMax, yMin + mv, 0];
+            const axes = [xMin, yMin + mv, zDepth, xMax, yMin + mv, zDepth];
             const [ticks, labels] = this.GetTickLinesAndLabels(
                 xMin,
                 xMax,
                 ViewSide.Bottom,
-                pixel2world
+                pixel2world,
+                zoomScale
             );
             const back_lines: number[] =
                 this.GetBacgroundTriangleLinesHorizontal(
@@ -522,12 +528,13 @@ export default class Axes2DLayer extends Layer<Axes2DLayerProps<unknown>> {
 
         //- TOP RULER ----------------------------------------
         if (this.props.isTopRuler) {
-            const axes = [xMin, yMax - mv, 0, xMax, yMax - mv, 0];
+            const axes = [xMin, yMax - mv, zDepth, xMax, yMax - mv, zDepth];
             const [ticks, labels] = this.GetTickLinesAndLabels(
                 xMin,
                 xMax,
                 ViewSide.Top,
-                pixel2world
+                pixel2world,
+                zoomScale
             );
 
             const back_lines = this.GetBacgroundTriangleLinesHorizontal(
@@ -547,12 +554,13 @@ export default class Axes2DLayer extends Layer<Axes2DLayerProps<unknown>> {
         if (this.props.isLeftRuler) {
             const ymin = this.props.isBottomRuler ? yMin + mv : yMin;
             const ymax = this.props.isTopRuler ? yMax - mv : yMax;
-            const axes = [xMin + mh, ymin, 0, xMin + mh, ymax, 0];
+            const axes = [xMin + mh, ymin, zDepth, xMin + mh, ymax, zDepth];
             const [ticks, labels] = this.GetTickLinesAndLabels(
                 ymin,
                 yMax - mv,
                 ViewSide.Left,
-                pixel2world
+                pixel2world,
+                zoomScale
             );
             const back_lines = this.GetBacgroundTriangleLinesVertical(
                 ymin,
@@ -571,12 +579,13 @@ export default class Axes2DLayer extends Layer<Axes2DLayerProps<unknown>> {
         if (this.props.isRightRuler) {
             const ymin = this.props.isBottomRuler ? yMin + mv : yMin;
             const ymax = this.props.isTopRuler ? yMax - mv : yMax;
-            const axes = [xMax - mh, ymin, 0, xMax - mh, ymax, 0];
+            const axes = [xMax - mh, ymin, zDepth, xMax - mh, ymax, zDepth];
             const [ticks, labels] = this.GetTickLinesAndLabels(
                 ymin,
                 ymax,
                 ViewSide.Right,
-                pixel2world
+                pixel2world,
+                zoomScale
             );
 
             const back_lines = this.GetBacgroundTriangleLinesVertical(
