@@ -1,4 +1,4 @@
-import React from "react";
+import React, { SyntheticEvent } from "react";
 import { styled } from "@mui/material/styles";
 import { ViewsType, useHoverInfo } from "../../components/Map";
 import SubsurfaceViewer from "../../SubsurfaceViewer";
@@ -427,7 +427,7 @@ export const ScaleZ: ComponentStory<typeof SubsurfaceViewer> = (args) => {
 
 ScaleZ.args = {
     id: "ScaleZ",
-    layers: [axes_hugin, meshMapLayerPng, north_arrow_layer],
+    layers: [axes_hugin, meshMapLayerPng, wellsLayer, north_arrow_layer],
     bounds: [432150, 6475800, 439400, 6481500] as NumberQuad,
 
     views: {
@@ -464,9 +464,9 @@ export const ResetCameraProperty: ComponentStory<typeof SubsurfaceViewer> = (
     const [home, setHome] = React.useState<number>(0);
     const [camera, setCamera] = React.useState({
         rotationOrbit: 0,
-        rotationX: 89,
-        target: [435775, 6478650, -1750],
-        zoom: -3.5109619192773796,
+        rotationX: 45,
+        target: [435775, 6477650, -1750],
+        zoom: -3.8,
     });
 
     const handleChange1 = () => {
@@ -488,7 +488,7 @@ export const ResetCameraProperty: ComponentStory<typeof SubsurfaceViewer> = (
             <div className={classes.main}>
                 <SubsurfaceViewer {...props} />
             </div>
-            <button onClick={handleChange1}> Reset Camera </button>
+            <button onClick={handleChange1}> Reset Camera to bounds</button>
             <button onClick={handleChange2}> Change Camera </button>
         </Root>
     );
@@ -1033,9 +1033,12 @@ export const BreakpointColorMap: ComponentStory<typeof SubsurfaceViewer> = (
         layers: [layer],
     };
 
-    const handleChange = React.useCallback((_event, value) => {
-        setBreakpoint(value / 100);
-    }, []);
+    const handleChange = React.useCallback(
+        (_event: Event | SyntheticEvent, value: number | number[]) => {
+            setBreakpoint((value as number) / 100);
+        },
+        []
+    );
 
     return (
         <Root>
@@ -1096,9 +1099,12 @@ export const ColorMapRange: ComponentStory<typeof SubsurfaceViewer> = (
         layers: [layer],
     };
 
-    const handleChange = React.useCallback((_event, value) => {
-        setColorMapUpper(value);
-    }, []);
+    const handleChange = React.useCallback(
+        (_event: unknown, value: number | number[]) => {
+            setColorMapUpper(value as number);
+        },
+        []
+    );
 
     return (
         <Root>
@@ -1152,26 +1158,44 @@ const MapLayerColorSelectorTemplate: ComponentStory<typeof SubsurfaceViewer> = (
     const [isNearest, setIsNearest] = React.useState(false);
 
     // user defined breakpoint(domain)
-    const userDefinedBreakPoint = React.useCallback((data) => {
-        if (data) setBreakPoint(data.colorArray);
-    }, []);
+    const userDefinedBreakPoint = React.useCallback(
+        (data: { colorArray: React.SetStateAction<undefined> }) => {
+            if (data) setBreakPoint(data.colorArray);
+        },
+        []
+    );
 
     // Get color name from color selector
-    const colorNameFromSelector = React.useCallback((data) => {
-        setColorName(data);
-    }, []);
+    const colorNameFromSelector = React.useCallback(
+        (data: React.SetStateAction<string>) => {
+            setColorName(data);
+        },
+        []
+    );
 
     // user defined range
-    const userDefinedRange = React.useCallback((data) => {
-        if (data.range) setRange(data.range);
-        setAuto(data.isAuto);
-    }, []);
+    const userDefinedRange = React.useCallback(
+        (data: {
+            range: React.SetStateAction<undefined>;
+            isAuto: React.SetStateAction<undefined>;
+        }) => {
+            if (data.range) setRange(data.range);
+            setAuto(data.isAuto);
+        },
+        []
+    );
 
     // Get interpolation method from color selector to layer
-    const getInterpolateMethod = React.useCallback((data) => {
-        setIsLog(data.isLog);
-        setIsNearest(data.isNearest);
-    }, []);
+    const getInterpolateMethod = React.useCallback(
+        (data: {
+            isLog: boolean | ((prevState: boolean) => boolean);
+            isNearest: boolean | ((prevState: boolean) => boolean);
+        }) => {
+            setIsLog(data.isLog);
+            setIsNearest(data.isNearest);
+        },
+        []
+    );
 
     // color map function
     const colorMapFunc = React.useCallback(() => {
